requirements.txt
asn1crypto==0.24.0
astroid==1.6.1
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==2.0.0
braintree==3.42.0
certifi==2017.7.27.1
cffi==1.11.4
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
diff-match-patch==20121119
Django==2.0.3
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-excel==0.0.6
django-extensions==1.7.9
django-filter==1.0.2
django-import-export==0.5.1
django-markdown-deux==1.0.5
django-numpy==1.0.1
django-pagedown==0.1.3
django-pandas==0.4.1
django-payments==0.12.1
django-paypal==0.5.0
django-widget-tweaks==1.4.1
djangorestframework==3.8.1
docutils==0.13.1
docx2txt==0.6
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.6
gunicorn==19.7.1
html5lib==0.999999999
hyperlink==17.3.1
idna==2.5
incremental==17.5.0
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.11
markdown2==2.3.4
matplotlib==2.1.2
mccabe==0.6.1
Mezzanine==4.2.3
nltk==3.2.5
numpy==1.12.0
oauthlib==2.0.2
olefile==0.44
pandas==0.19.2
pbr==4.0.1
pdfrw==0.4
Pillow==4.2.1
psycopg2==2.7.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyBarcode==0.7
pycparser==2.18
PyDispatcher==2.0.5
pyexcel==0.4.5
pyexcel-io==0.3.3
pyexcel-webio==0.0.11
PyJWT==1.6.0
pylint==1.8.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
pytz==2016.10
queuelib==1.4.2
requests==2.18.2
requests-oauthlib==0.8.0
scikit-learn==0.19.1
service-identity==17.0.0
six==1.11.0
stevedore==1.28.0
stripe==1.79.1
suds-jurko==0.6
tablib==0.11.4
texttable==0.8.8
tweepy==3.5.0
tzlocal==1.4
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
webencodings==0.5.1
wrapt==1.10.11
xlwt==1.2.0
xmltodict==0.11.0
zope.interface==4.4.3
